requirements.txt
streamlit==1.31.0
pandas==2.1.1
python-dotenv==1.0.1
google-generativeai==0.3.2
numpy>=1.26
